Try to demonstrate problem with fast-import's truncation
When fast-import writes objects, it may detect only after writing an
object that this object was already inserted into the object database.
In that case, fast-import rolls back what it has written last,
essentially "unwriting" the duplicated object.

It has been reported that this truncation can fail, with

	fatal: cannot truncate pack to skip duplicate: Invalid argument

This (failed) attempt tries to replicate that behavior with a toy
import.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>