string-list: introduce a function to delete items from sorted lists
There is already a function that deletes items from unsorted lists. Due
to the unsorted nature, that function simply reorders the items in the
interest of optimal time complexity (constant time).

However, if the list is sorted (and the intention is to keep it sorted),
then the operation needs to move the remaining items, which is
necessarily more costly (linear time).

Still, it is a useful operation to have. As it requires careful
consideration whether a different strategy would be more performant
(e.g. by using the unsorted version of the function for a number of
items and simply re-sort afterwards), we put a little explanation of the
ramifications on top of the declaration in the header file.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>