mingw: remember the current directory
We already need to know the length of the current directory's path, but
so far, we had no use for the actual path.

To prepare for being able to convert absolute paths *without* a drive
prefix to paths *with* that prefix, we will need to know the current
drive, though, so let's just keep a record of the current directory at
all times.

Note: this is not thread-safe, intentionally so: we should avoid *at all
cost* to change the current working directory in a multi-threaded
operation. It would be much too fragile to do so.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>