fixup! stash: allow pathspecs in the no verb form
The convention is to move obsolete scripts into contrib/examples/. While
this convention is questionable (it is intended for those scripts to serve
as examples how to script Git, but there is *no guarantee whatsoever* that
they still work, or still follow coding style), it is useful in the case
of git-stash.sh:

When rebasing the rename, it will not conflict with changes made to
git-stash.sh in the meantime, but patch contrib/examples/git-stash.sh.
These changes can then be used as an indicator how the rebased
builtin/stash.c needs to be adjusted to follow upstream's changes.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>