fixup! stash: implement builtin stash
Before calling refresh_index(), ensure that the cache has actually been
read.

It is a common pitfall to encounter: refresh_index() succeeds happily,
having refreshed all (exactly zero) index items. That function simply does
not consider it part of its job to read the .git/index file. So that is
the job of the caller, to ensure that the index has been populated
appropriately (the rationale: it could be an in-memory index, or an index
read from elsewhere).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>