DONTMERGE: unpack-trees: preserve index extensions
Make git checkout (and other unpack_tree operations) preserve the
untracked cache. This is valuable for two reasons:

1. Often, an unpack_tree operation will not touch large parts of the
working tree, and thus most of the untracked cache will continue to be
valid.

2. Even if the untracked cache were entirely invalidated by such an
operation, the user has signaled their intention to have such a cache,
and we don't want to throw it away.

[jes: backed out the watchman-specific parts]

[DONTMERGE: in <2c0be7c0a8ce4f5ea49dfaf3f8511bc0@exmbdft7.ad.twosigma.com>,
the original author says "The tests do not correctly check that the
post-checkout untracked cache is still valid after a checkout.  For
example, let's say that previously, the directory foo was entirely
untracked (but it contained a file bar), but after the checkout,
there is a file foo/baz.  Does the untracked cache need to get
updated?  ... But I would strongly suggest that the test be improved
before this code is merged."]

Signed-off-by: David Turner <dturner@twopensource.com>
Signed-off-by: Ben Peart <benpeart@microsoft.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>